fix int8 and nf4 cpu inference
